refactor: template 파일 layout 폴더로 이동
